requirements.txt
scipy
crowdai_api
dm-sonnet==1.35
minerl==0.2.9
coloredlogs
matplotlib
seaborn
numpy==1.17.2
opencv-python
pyro4
tensorflow-gpu==1.14.0
tensorflow-probability==0.7.0
wrapt==1.11.1
aicrowd-repo2docker==0.9.1
